adding tests

+++ trunk/src/hpbtc/protocol/network/TrackerConnection.java
@@ -38,10 +38,10 @@ public class TrackerConnection {
     private long lastCheck;
     private byte[] infoHash;
     private byte[] pid;
-    private long port;
+    private int port;
     private List<LinkedList<String>> trackers;
     
-    public TrackerConnection(byte[] infoHash, byte[] pid, long port, List<LinkedList<String>> trackers) {
+    public TrackerConnection(byte[] infoHash, byte[] pid, int port, List<LinkedList<String>> trackers) {
         this.infoHash = infoHash;
         this.pid = pid;
         this.port = port;
@@ -143,7 +143,7 @@ public class TrackerConnection {
                 String id = beid.getValue();
                 if (!Arrays.equals(pid, beid.getBytes())) {
                     Peer p = new Peer(((BencodedString) d.get("ip")).getValue(),
-                            ((BencodedInteger) d.get("port")).getValue(), id);
+                            ((BencodedInteger) d.get("port")).getValue().intValue(), id);
                     peers.add(p);
                 }
             }
